Move code to SoundcloudTrack class
