+++ pmg/models/__init__.py
@@ -2,4 +2,4 @@ from .users import *
 from .resources import *
 from .emails import *
 from .pages import *
-from .soundcloud import *
+from .soundcloud_track import *
